server/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
PyJWT
passlib[bcrypt]
slowapi
pytest
httpx
